[ODIN-22] Reorganized import and remove white lines
