+++ src/app/models/SceneSettings.ts
@@ -1,5 +1,5 @@
 export class SceneSettings {
     backgroundColor: string;
 
-    constructor(){}
+    constructor() { }
 }
