+++ src/app/models/logModels/logLevel.ts
@@ -1,9 +1,9 @@
 export enum LogLevel {
-    All = 0,
-    Debug = 1,
-    Info = 2,
-    Warn = 3,
-    Error = 4,
-    Fatal = 5,
-    Off = 6
-  }
+  All = 0,
+  Debug = 1,
+  Info = 2,
+  Warn = 3,
+  Error = 4,
+  Fatal = 5,
+  Off = 6
+}
