+++ src/app/models/logModels/meshLog.ts
@@ -1,8 +1,5 @@
 import { Mesh } from 'babylonjs';
 
 export class MeshLog {
-    
-    constructor(m: Mesh){
-        
-    }
+    constructor(m: Mesh) { }
 }
