+++ src/app/pipes/precision.pipe.ts
@@ -9,5 +9,4 @@ export class PrecisionPipe implements PipeTransform {
   transform(value: number): number {
     return Utils.precision(value, 3);
   }
-
 }
