+++ src/app/services/canvas-helper.service.ts
@@ -1,13 +1,13 @@
-import { Utils } from './../engine/Utils/Utils';
 import { Injectable } from '@angular/core';
 import { select, Store } from '@ngrx/store';
-import { Light, LightGizmo, Mesh, PickingInfo, ShadowLight, GizmoManager, Vector3, BoundingInfo, Color3 } from 'babylonjs';
-import { take } from 'rxjs/operators';
-import { AppState } from '../app.reducer';
-import { EngineService } from '../engine/engine.service';
-import { clearSelection, oneSelection } from '../engine/engine.action';
-import { Container } from '../engine/common/Container';
+import { BoundingInfo, Color3, GizmoManager, Light, LightGizmo, Mesh, PickingInfo, ShadowLight, Vector3 } from 'babylonjs';
 import { BoundingBox } from 'babylonjs/Culling/boundingBox';
+import { take } from 'rxjs/operators';
+import { Container } from '../engine/common/Container';
+import { clearSelection, oneSelection } from '../store/actions/engine.actions';
+import { AppState } from '../store/reducers/app.reducer';
+import { Utils } from './../engine/Utils/Utils';
+import { EngineService } from './engine.service';
 
 @Injectable({ providedIn: 'root' })
 export class CanvasHelperService {
