+++ src/app/services/index.service.ts
@@ -1,4 +1,7 @@
-export { WindowService as WindowRefService } from './window.service';
-export { EngineService } from '../engine/engine.service';
 export { AppService } from './app.service';
 export { CanvasHelperService } from './canvas-helper.service';
+export { EngineService } from './engine.service';
+export { LibraryService } from './library.service';
+export { LogService } from './log.service';
+export { SessionService } from './session.service';
+export { WindowService as WindowRefService } from './window.service';
