+++ src/app/services/service.module.ts
@@ -1,18 +1,14 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { WindowService } from './window.service';
-import { EngineService } from '../engine/engine.service';
-import { LogService } from './log.service';
-import { AppService } from './app.service';
-import { CanvasHelperService } from './canvas-helper.service';
-import { SessionService } from './session.service';
+import { NgModule } from '@angular/core';
+import { AppService, CanvasHelperService, EngineService, LogService, SessionService, WindowRefService } from './index.service';
+
 
 @NgModule({
     imports: [CommonModule],
     exports: [],
     declarations: [],
     providers: [
-        WindowService,
+        WindowRefService,
         EngineService,
         LogService,
         AppService,
