+++ src/app/shared/alert/alert.component.html
@@ -1,5 +1,4 @@
-<div style="padding: 2px;" [ngClass]="{'inline': inline}" 
-[ngClass]="{'alertType-success': type == 0,
+<div style="padding: 2px;" [ngClass]="{'inline': inline}" [ngClass]="{'alertType-success': type == 0,
             'alertType-info': type == 1,
             'alertType-warning':type == 2,
             'alertType-error':type == 3
