+++ src/app/shared/alert/alert.component.scss
@@ -1,15 +1,15 @@
-.alertType-error{
+.alertType-error {
     background-color: crimson;
 }
 
-.alertType-success{
+.alertType-success {
     background-color: seagreen;
 }
-.alertType-info{
+.alertType-info {
     background-color: lightseagreen;
 }
 
-.alertType-warninf{
+.alertType-warninf {
     background-color: goldenrod;
 }
