+++ src/app/store/actions/session.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { User } from '../models/User';
+import { User } from '../../models/User';
 
 export const login = createAction('[AUTHENTICATION] LOGIN', props<{ user: User }>());
 export const signIn = createAction('[AUTHENTICATION] SIGNIN', props<{ user: User }>());
