+++ src/app/store/reducers/app.reducer.ts
@@ -1,7 +1,7 @@
 import { ActionReducerMap } from "@ngrx/store";
-import * as engine from './engine/engine.reducer'
-import * as session from './services/session.reducer'
-import * as ui from './ui/ui.reducer'
+import * as engine from './engine.reducer';
+import * as session from './session.reducer';
+import * as ui from './ui.reducer';
 
 export interface AppState {
     engine: engine.State;
