+++ src/app/ui/about-odin/about-odin.component.html
@@ -2,6 +2,7 @@
   <div>
     <a href="https://github.com/VallholRT/Odin" target="_blank"></a>
     <h3>Version 0.1.1 | Sister Sin - In It for Life | 23/03/2021 </h3>
-    <p>Odin is a Broadcast web app designer, createad by <a href="https://www.sparanzza.com" target="_blank">Aurelio Muñoz Hijano</a></p>
+    <p>Odin is a Broadcast web app designer, createad by <a href="https://www.sparanzza.com" target="_blank">Aurelio
+        Muñoz Hijano</a></p>
   </div>
 </popup-window>
