+++ src/app/ui/camera-panel/camera-panel.component.scss
@@ -2,23 +2,24 @@ input[type="checkbox"] {
 	-webkit-appearance: none;
 	background-color: #fafafa;
 	border: 1px solid #cacece;
-	box-shadow: 0 1px 2px rgba(0,0,0,0.05), inset 0px -15px 10px -12px rgba(0,0,0,0.05);
+	box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05), inset 0px -15px 10px -12px rgba(0, 0, 0, 0.05);
 	padding: 9px;
 	border-radius: 3px;
 	display: inline-block;
 	position: relative;
-    width: 16px;
+	width: 16px;
 }
 
 input[type="checkbox"]:checked {
 	background-color: #e9ecee;
 	border: 1px solid #adb8c0;
-	box-shadow: 0 1px 2px rgba(0,0,0,0.05), inset 0px -15px 10px -12px rgba(0,0,0,0.05), inset 15px 10px -12px rgba(255,255,255,0.1);
+	box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05), inset 0px -15px 10px -12px rgba(0, 0, 0, 0.05),
+		inset 15px 10px -12px rgba(255, 255, 255, 0.1);
 	color: #99a1a7;
 }
 
 input[type="checkbox"]:checked:after {
-	content: '\2714';
+	content: "\2714";
 	font-size: 14px;
 	position: absolute;
 	top: 0px;
