+++ src/app/ui/console-debug/console-debug.component.scss
@@ -5,7 +5,7 @@
     background: #111;
 }
 
-.inner-text-console{
+.inner-text-console {
     margin: 0px;
     font-size: 12px;
     text-align: left;
@@ -30,6 +30,6 @@
 }
 
 ::-webkit-scrollbar {
-    width: 0px;  /* Remove scrollbar space */
-    background: transparent;  /* Optional: just make scrollbar invisible */
+    width: 0px; /* Remove scrollbar space */
+    background: transparent; /* Optional: just make scrollbar invisible */
 }
