+++ src/app/ui/content-panel/content-panel.component.scss
@@ -18,7 +18,6 @@ input {
 }
 
 label {
-
     color: white;
     margin: 0px 5px;
     width: 5px;
@@ -30,10 +29,6 @@ label {
     padding: 2px;
     margin: 2px;
     border: var(--main-menu-border);
-}
-
-.col-transform-menu {
-
 }
 
 .title-section-menu {
