+++ src/app/ui/scene-settings/scene-settings.component.scss
