+++ src/app/ui/text-panel/text-panel.component.scss
@@ -7,7 +7,7 @@
     right: 0px;
 }
 
-.title-section-menu{
+.title-section-menu {
     font-size: 10px;
     color: var(--console-debug-color);
 }
