+++ src/app/ui/tree-node/tree-node.component.scss
@@ -180,6 +180,6 @@ button {
   height: 18px;
 }
 
-.tree-node__tope-menu__group-vertical-btn{
+.tree-node__tope-menu__group-vertical-btn {
   width: min-content;
 }
