+++ src/app/ui/viewport/viewport.component.html
@@ -1,4 +1,3 @@
 <div class="engine-wrapper">
     <canvas #rendererCanvas id="renderCanvas"></canvas>
 </div>
-
